requirements.txt
easydict==1.11
insightface==0.7.3
numpy==1.26.3
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
pandas==2.2.0
Pillow==10.2.0
pytorch_lightning==2.1.3
PyYAML==6.0.1
scikit_learn==1.3.2
timm==0.9.12
torch==2.1.2
torchmetrics==1.2.1
torchsampler==0.1.2
torchvision==0.16.2
